chore: Converted context-manager unit tests to node:test (#2508)
